move cs to deployment package
